@@ -3,17 +3,13 @@
   "version": "2.0.1",
   "description": "Play the game \"Set\"",
   "main": "lib/index.js",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/hildjj/set-game.git"
   },
   "scripts": {
-    "test": "ava",
-    "coverage": "nyc --reporter=html npm test",
-    "lint": "standard",
-    "release": "npm version patch && git push --follow-tags && npm publish",
-    "dev": "npm run coverage && light-server -w 'lib/*.js test/*.js # npm run coverage # reload' -o coverage",
-    "coveralls": "nyc report --reporter=text-lcov | coveralls"
+    "test": "c8 ava"
   },
   "author": {
     "name": "Joe Hildebrand",
@@ -24,20 +20,32 @@
     "bin": "bin"
   },
   "dependencies": {
-    "chalk": "^4.0.0",
-    "js-combinatorics": "^0.5.2",
-    "knuth-shuffle": "^1.0.1",
-    "pug": "^2.0.4",
-    "yargs": "^15.3.1"
+    "chalk": "^5.3.0",
+    "js-combinatorics": "^2.1.1",
+    "knuth-shuffle": "^1.0.8",
+    "pug": "^3.0.2",
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "ava": "^3.8.2",
-    "light-server": "github:hildjj/light-server#update-dependencies-2020-05-11",
-    "nyc": "^15.0.1",
-    "pug-cli": "^1.0.0-alpha6",
-    "standard": "^14.3.4"
+    "@cto.af/eslint-config": "2.0.3",
+    "ava": "^5.3.1",
+    "c8": "8.0.0",
+    "eslint": "8.44.0",
+    "eslint-plugin-jsdoc": "46.4.3",
+    "eslint-plugin-n": "16.0.1",
+    "js-xiterable": "0.2.2"
+  },
+  "pnpm": {
+    "overrides": {
+      "semver": ">=7.5.3",
+      "word-wrap": "npm:@aashutoshrathi/word-wrap"
+    }
+  },
+  "overrides": {
+    "semver": ">=7.5.3",
+    "word-wrap": "npm:@aashutoshrathi/word-wrap"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=16"
   }
 }
