@@ -1,39 +1,41 @@
-"use strict"
+import {fileURLToPath} from 'url'
+import path from 'path'
+import pug from 'pug'
+import util from 'util'
 
-const path = require('path')
-const util = require('util')
-const pug = require('pug')
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 
 /**
  * A Formatted class uses one or more Pug templates to translate into
  * different formats.
  */
-class Formatted {
-  // cache
+export class Formatted {
+  // Cache
   static #compiled = {}
 
   /**
-   * The directory to find Pug template files in
+   * The directory to find Pug template files in.
    * @static
    */
   static TEMPLATE_DIR = path.join(__dirname, '..', 'templates')
 
   /**
-   * Reset the cache
+   * Reset the cache.
    * @static
    */
-  static reset () {
+  static reset() {
     Formatted.#compiled = {}
   }
 
   /**
-   * Execute a template of the give output type
+   * Execute a template of the give output type.
    *
-   * @param {string} outputType - e.g. "text" or "html"
-   * @param {object} [opts={}] - util.inspect options, or anything else you want
-   * @returns {string} The rendered template
+   * @param {string} outputType E.g. "text" or "html".
+   * @param {object} [opts={}] Options for util.inspect, or anything else.
+   * @returns {string} The rendered template.
    */
-  template (outputType, opts={}) {
+  template(outputType, opts = {}) {
     const key = `${this.constructor.name.toLowerCase()}-${outputType}`
     let js = Formatted.#compiled[key]
     if (!js) {
@@ -45,35 +47,33 @@ class Formatted {
   }
 
   /**
-   * Render the HTML template for this object
+   * Render the HTML template for this object.
    *
-   * @param {object} [opts={}] - util.inspect options, or anything else you want
-   * @returns {string} The rendered template
+   * @param {object} [opts={}] Options for util.inspect, or anything else.
+   * @returns {string} The rendered template.
    */
-  html (opts) {
+  html(opts) {
     return this.template('html', opts)
   }
 
   /**
-   * Render the text template for this object
+   * Render the text template for this object.
    *
-   * @param {object} [opts={}] - util.inspect options, or anything else you want
-   * @returns {string} The rendered template
+   * @param {object} [opts={}] Options for util.inspect.
+   * @returns {string} The rendered template.
    */
-  text (opts) {
+  text(opts) {
     return this.template('text', opts)
   }
 
   /**
-   * Use the text template for visualizing the object in the node REPL
+   * Use the text template for visualizing the object in the node REPL.
    *
-   * @param {number} depth - depth to descend
-   * @param {object} opts - see util.inspect for options
-   * @returns
+   * @param {number} depth Dpth to descend.
+   * @param {object} opts See util.inspect for options.
+   * @returns {string} Potentially-colorful text.
    */
-  [util.inspect.custom] (depth, opts) {
+  [util.inspect.custom](depth, opts) {
     return this.text(opts)
   }
 }
-
-module.exports = Formatted
