@@ -1,8 +1,11 @@
-'use strict'
+import {Formatted} from '../lib/formatted.js'
+import {fileURLToPath} from 'url'
+import path from 'path'
+import test from 'ava'
+import util from 'util'
 
-const util = require('util')
-const test = require('ava')
-const Formatted = require('../lib/formatted')
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 
 test.before(t => {
   t.context.TEMPLATE_DIR = Formatted.TEMPLATE_DIR
@@ -17,7 +20,7 @@ test('create', t => {
   const f = new Formatted()
   t.is(f.html(), '\n<div class="formatted"></div>')
   t.is(f.text(), 'Formatted')
-  t.is(f.text(), 'Formatted') // check the cache
+  t.is(f.text(), 'Formatted') // Check the cache
   Formatted.reset()
   t.is(util.inspect(f), 'Formatted')
 })
